Doruk - Ekleme / Güncelleme

+++ user-service/src/main/java/com/ikolay/controller/UserController.java
@@ -75,7 +75,7 @@ public class UserController {
     }
 
     @GetMapping("/monthlyemployeesalary/{companyId}")
-    ResponseEntity<Long> findTotalEmployeeSalary(@PathVariable Long companyId){
+    ResponseEntity<Double> findTotalEmployeeSalary(@PathVariable Long companyId){
         System.out.println(userService.findTotalEmployeeSalary(companyId));
         return ResponseEntity.ok(userService.findTotalEmployeeSalary(companyId));
     }
